Reworking "splitters". Part 1 of at least 2, and likely 3.
Symmetric splitters are splitters that use the same text string (which
may be the empty string) before and after the selected content. Paired
splitters (coming soon) are represented as HTML tag pairs with ID
attribute strings that end in either "-begin" or "-end".

27 tests, 31 assertions, 0 failures, 0 errors, 0 skips
Coverage: 84 / 84 LOC (100.0%) covered
RuboCop: 5 files inspected, no offenses detected
Flay: Total score 0
Flog: Total 69.3; method average 3.5; max 5.5 (Prolog::Services::ReplaceContent#initialize)
Reek: 0 warnings
